Cleanup. Follow best practices. Use proper TypeScript typing.
